create items table, migration

+++ src/migrations/list/create-sockets-table_1570627775991.sql
@@ -1,5 +1,5 @@
 create table sockets (
-  id int not null auto_increment,
+  id smallint unsigned not null auto_increment,
   socket_group tinyint not null,
   attr varchar(5) not null,
   colour varchar(5) not null,
