+++ src/migrations/list/create_leagues_table_1570626875496.sql
@@ -1,5 +1,5 @@
 create table leagues (
-  id int not null auto_increment,
+  id tinyint(3) unsigned not null auto_increment,
   name varchar(255) not null unique,
   primary key (id)
 );
