+++ src/migrations/list/create_subcategory_table_1570627116706.sql
@@ -1,5 +1,5 @@
 create table subcategories (
-  id int not null auto_increment,
+  id smallint unsigned not null auto_increment,
   name varchar(255) not null unique,
   primary key (id)
 );
