+++ src/migrations/list/create_typeline_table_1570626400960.sql
@@ -1,5 +1,5 @@
 create table typelines (
-  id int not null auto_increment,
+  id mediumint unsigned not null auto_increment,
   name varchar(255) not null unique,
   primary key (id)
 );
